BFD-1744: migrate dme claims new schema (#1105)
* merge changes from bfd1536; add new sql script to create SNF claims int new schema format

* SNF claims data migration script

* cleanup comments (spelling, punctuation, etc.)

* revert numeric(10,2) back to numeric(12,2); minor reformatting

* fix branch due to bad merges; this branch is built on branches not yet in master

* changes for SNF so trotter can build branch

* remove commented out snf_claims from 'future' bigint

* remove support for table views

* remove support for table views

* minor formatting of if block

* fix some deprecated code items

* minor tweak of numeric value to match other test values

* add Javadoc info in a couple of places

* rename flyway migration script due to master changes

* add JavaDoc comments; minor addition of if-then-else brackets

* fix merge issues

* revert changes that included brackets

* fix merge from bfd1595 by removing view creation

* annother try to ensure V63 migration script matches bfd1595 script

* Revert SNF claims to use original table structure

* Revert tests to use original SNF layout

* minor mod to selectJavaDataType to simplify handling NUM entries

* make sure REV_CNTR_NDC_QTY supports decimal values

* flyway script to fix new SNF claims tables and migrate data

* revert most integer columns back to numeric

* revert some of the SNF table columns to NUMERIC

* chg a couple of SNF rows to use NUM

* changes for REV_CNTR_NDC_QTY test data to include decimal value

* added a comment to NDC qty

* minor excel formatting

* reconcile V66 sql with sql in bfd1596

* merge bfd1596 into this branch

* fix spelling errors in comments

* flyway scripts to migrate DME claims to new schema

* fix dmerc_line_mtus_cnt to include precision and scale

+++ apps/bfd-pipeline/bfd-pipeline-ccw-rif/src/test/java/gov/cms/bfd/pipeline/ccw/rif/extract/RifFilesProcessorTest.java
@@ -604,7 +604,7 @@ public final class RifFilesProcessorTest {
     assertEquals(claimGroup.getBeneficiaryId(), rifRecordEvent.getBeneficiaryId());
     assertEquals(567834L, claimGroup.getBeneficiaryId());
     assertEquals(777777777L, claimGroup.getClaimId());
-    assertEquals(BigDecimal.valueOf(900L), claimGroup.getClaimGroupId());
+    assertEquals(900L, claimGroup.getClaimGroupId());
     assertEquals('V', claimGroup.getNearLineRecordIdCode());
     assertEquals("20", claimGroup.getClaimTypeCode());
     assertEquals(LocalDate.of(2013, 12, 01), claimGroup.getDateFrom());
